requirements_prod.txt
# TruthTracker API Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# ML/AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
facenet-pytorch>=2.5.3
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Explainability
lime>=0.2.0.1
shap>=0.43.0
pytorch-grad-cam>=1.5.0

# Optional: XGBoost for ensemble
xgboost>=2.0.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
